Added via PC look at Replit for website Final Wesbite
Had some troubles pushing this via the use of Replit as the Github kept crashing. there is meant to be a commit 5 days ago for the API timer. This commit is for the login page and the the content related changes. Still use the website on replit for the best experience this is just for